+++ esp-hal/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- `InterruptHandler` no longer implements `PartialEq`, `Eq` or `Hash`. (#3650)
 
 ## [v1.0.0-beta.1] - 2025-06-03